Fixing a lot of ugly hacks and integrating interface for new functionalities
A lot of functionality (like re-publishing logged designators) is now part the actual C++ beliefstate logger. Also, support for new functionalities is introduced (like setting the meta data for an experiment). Parts of the hooks that are called during task execution are already replaced by the more CRAM-like `define-hook` callbacks. More to come.