Interface cleanup and ensuring that designators are annotated to the correct nodes
Designators (and objects/locations in particular) were sometimes ending up on the wrong node context in the logging system. This is fixed now as new functionalities in the logging system are employed that can annotate the designators to a specific node ID rather than the currently active one. This solves a lot of inconsistency problems. Also, cleanup of a few functions.